centralizing the bitinfo calculating so I can change it in the future changing some types to be more explicit

+++ pinochle_test.go
@@ -75,7 +75,7 @@ func (t *testSuite) TestDeal() {
 
 func (t *testSuite) TestSmallHand() {
 	hand := NewSmallHand()
-	for card := AS; card < AllCards; card++ {
+	for card := AS; int8(card) < AllCards; card++ {
 		t.False(hand.Contains(card))
 	}
 	t.False(hand.Contains(JD))
